VS code/tidligere forsøk/kyst-monitor-v3/requirements.txt
# API & HTTP
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Auth & Credentials
python-dotenv==1.0.0
PyJWT==2.10.1

# Database
sqlalchemy==2.0.23

# Data Processing
# pandas==2.1.3
# numpy==1.26.2

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Scheduling
APScheduler==3.10.4

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
